Fixes to xPDOObject -- Muchos thanks to opengeek
	modified:   xpdo/om/pgsql/xpdoobject.class.php
	modified:   xpdo/om/xpdodriver.class.php
	modified:   xpdo/om/xpdoobject.class.php

+++ test/model/sample/mysql/person.map.inc.php
@@ -90,7 +90,7 @@ $xpdo_meta_map['Person']= array (
     ),
     'security_level' => 
     array (
-      'dbtype' => 'smallint',
+      'dbtype' => 'tinyint',
       'precision' => '4',
       'phptype' => 'integer',
       'null' => false,
